# Optimise link expansion to remove the unnecessarily database queries