For dependees:
  Use query to encapsulate link_expansion logic and retrieve all the
  target_content_ids that match those rules. Fetch all the
  web_content_items in that set in one query, which removes the need to
  iteratively fetch the next level.

For dependents:
  Limit query to search for dependents that are in the recursive set.
  Fetch all web_content_items in one query
  Cache the parent lookup since it will always be the same

Do not include non renderable formats when fetching web content items

Roughly 61 seconds savings:

Before: bench/downstream_payload.rb
Many reverse dependencies: 4bda0be5-3e65-4cc1-850c-0541e95a40ca
.......... 41.350000   2.190000  43.540000 ( 71.211850)
queries: 41465

Many forward dependencies: 5f4f1e91-7631-11e4-a3cb-005056011aef
..........  2.020000   0.080000   2.100000 (  5.351877)
queries: 220

No dependencies: 23e13323-874f-4210-aad3-ca6ade9206f8
..........  0.080000   0.010000   0.090000 (  0.117896)
queries: 100

Single link each way: 00012147-49f6-4e90-be1f-e50bb719f53d
..........  0.140000   0.010000   0.150000 (  0.264822)
queries: 160

After: bench/downstream_payload.rb
Many reverse dependencies: 4bda0be5-3e65-4cc1-850c-0541e95a40ca
..........  3.930000   0.170000   4.100000 (  8.843297)
queries: 165

Many forward dependencies: 5f4f1e91-7631-11e4-a3cb-005056011aef
..........  2.160000   0.050000   2.210000 (  4.139760)
queries: 120

No dependencies: 23e13323-874f-4210-aad3-ca6ade9206f8
..........  0.070000   0.000000   0.070000 (  0.108529)
queries: 100

Single link each way: 00012147-49f6-4e90-be1f-e50bb719f53d
..........  0.100000   0.010000   0.110000 (  0.201108)
queries: 120

------------------------------------------------------------------------------

Before: bench/downstream_presenter.rb
Many reverse dependencies: 6667cce2-e809-4e21-ae09-cb0bdc1ddda3
..........
2.380000 0.200000 2.580000 ( 5.690258)
Many forward dependencies: 5f4f1e91-7631-11e4-a3cb-005056011aef
..........
1.990000 0.100000 2.090000 ( 5.768786)
No dependencies: 23e13323-874f-4210-aad3-ca6ade9206f8
..........
0.040000 0.010000 0.050000 ( 0.058350)
Single link each way: 00012147-49f6-4e90-be1f-e50bb719f53d
..........
0.100000 0.000000 0.100000 ( 0.195654)
Import export topic: 4bda0be5-3e65-4cc1-850c-0541e95a40ca
..........
42.060000 2.370000 44.430000 ( 72.970935)

After: bench/downstream_presenter.rb
Many reverse dependencies: 6667cce2-e809-4e21-ae09-cb0bdc1ddda3
  ..........
0.080000   0.010000   0.090000 (  0.138281)
  Many forward dependencies: 5f4f1e91-7631-11e4-a3cb-005056011aef
  ..........
2.110000   0.050000   2.160000 (  3.202435)
  No dependencies: 23e13323-874f-4210-aad3-ca6ade9206f8
  ..........
0.050000   0.010000   0.060000 (  0.089089)
  Single link each way: 00012147-49f6-4e90-be1f-e50bb719f53d
  ..........
0.080000   0.010000   0.090000 (  0.149320)
  Import export topic: 4bda0be5-3e65-4cc1-850c-0541e95a40ca
  ..........
3.760000   0.120000   3.880000 (  7.764834)